+++ src/components/Feeds/FeedListView.tsx
@@ -37,7 +37,8 @@ import { setSidebarActive } from "../../store/ui/actions";
 import { AddNodeProvider } from "../AddNode/context";
 import { DataTableToolbar, InfoIcon } from "../Common";
 import CreateFeed from "../CreateFeed/CreateFeed";
-import { CreateFeedProvider, PipelineProvider } from "../CreateFeed/context";
+import { CreateFeedProvider } from "../CreateFeed/context";
+import { PipelineProvider } from "../PipelinesCopy/context";
 import { ThemeContext } from "../DarkTheme/useTheme";
 import IconContainer from "../IconContainer";
 import WrapperConnect from "../Wrapper";
