--- src/components/Pipelines/ListCompute.tsx
@@ -1,55 +1,0 @@
-import { List, Avatar, Checkbox } from "antd";
-import { stringToColour } from "../CreateFeed/utils";
-
-const ListCompute = ({
-  computeList,
-  generalCompute,
-  dispatchFn,
-}: {
-  computeList: any[];
-  generalCompute?: string;
-  dispatchFn: (item: any) => void;
-}) => {
-  return (
-    <>
-      <List
-        itemLayout="horizontal"
-        dataSource={computeList && computeList.length > 0 ? computeList : []}
-        renderItem={(item: any) => {
-          return (
-            <List.Item>
-              <List.Item.Meta
-                avatar={
-                  <>
-                    <Checkbox
-                      style={{
-                        marginRight: "0.5em",
-                      }}
-                      checked={
-                        generalCompute && generalCompute === item.name
-                          ? true
-                          : false
-                      }
-                      onClick={() => {
-                        dispatchFn(item);
-                      }}
-                    />
-                    <Avatar
-                      style={{
-                        background: `${stringToColour(item.name)}`,
-                      }}
-                    />
-                  </>
-                }
-                title={item.name}
-                description={item.description}
-              />
-            </List.Item>
-          );
-        }}
-      />
-    </>
-  );
-};
-
-export default ListCompute;
