--- src/components/Pipelines/index.ts
@@ -1,7 +1,0 @@
-import NodeData from "./NodeData";
-import Tree from "./Tree";
-import UploadJson from "./UploadJson";
-import ConfigurationPage from "./ConfigurationPage";
-import GeneralCompute from "./GeneralCompute";
-
-export { NodeData, Tree, UploadJson, ConfigurationPage, GeneralCompute };
