+++ src/components/PipelinesCopy/Pipelines.css
@@ -1,0 +1,4 @@
+
+.ant-form-item-label label {
+   color: #73bcf7 !important;
+}
