--- src/components/PipelinesCopy/Pipelines.module.css
@@ -1,4 +1,0 @@
-
-	label {
-		color:#1fa7f8 !important;
-	}
