+++ src/components/PipelinesCopy/PipelinesComponent.tsx
@@ -1,7 +1,7 @@
 import { Pipeline } from "@fnndsc/chrisapi";
+import { Grid, GridItem } from "@patternfly/react-core";
 import CodeBlockComponent from "./CodeBlockComponent";
 import ComputeListForSingleCompute from "./ComputeListForSingleCompute";
-import { Grid, GridItem } from "@patternfly/react-core";
 import GeneralCompute from "./GeneralCompute";
 import TitleChange from "./TitleChange";
 import Tree from "./Tree";
@@ -15,21 +15,22 @@ function PipelinesComponent(props: OwnProps) {
   return (
     <>
       <Tree currentPipeline={pipeline} />
+
       <Grid
         style={{
-          marginTop: "2em",
+          marginTop: "3em",
         }}
         hasGutter={true}
       >
-        <GridItem span={4}>
+        <GridItem span={12} md={4} lg={5} xl={4}>
           <GeneralCompute />
         </GridItem>
 
-        <GridItem span={4}>
+        <GridItem span={12} md={6} lg={5} xl={4}>
           <ComputeListForSingleCompute currentPipeline={pipeline} />
         </GridItem>
 
-        <GridItem span={4}>
+        <GridItem span={12} md={2} lg={2} xl={4}>
           <TitleChange currentPipeline={pipeline} />
         </GridItem>
       </Grid>
